Pull CopyTerraformButton into own file for re-use
